Split composite constraints tests to separate acceptance test
