+++ enterprise/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/CompositeIndexAcceptanceTest.scala
@@ -243,66 +243,6 @@ class CompositeIndexAcceptanceTest extends ExecutionEngineFunSuite with NewPlann
     result should evaluateTo(List(Map("x" -> 1), Map("x" -> 3), Map("x" -> 5), Map("x" -> 7), Map("x" -> 9)))
   }
 
-  test("should be able to create and remove composite uniquness constraints") {
-    // When
-    executeWithCostPlannerAndInterpretedRuntimeOnly("CREATE CONSTRAINT ON (n:Person) ASSERT n.email IS UNIQUE")
-    executeWithCostPlannerAndInterpretedRuntimeOnly("CREATE CONSTRAINT ON (n:Person) ASSERT (n.firstname,n.lastname) IS UNIQUE")
-
-    // Then
-    graph should haveConstraints("UNIQUENESS:Person(email)", "UNIQUENESS:Person(firstname,lastname)")
-
-    // When
-    executeWithCostPlannerAndInterpretedRuntimeOnly("DROP CONSTRAINT ON (n:Person) ASSERT (n.firstname,n.lastname) IS UNIQUE")
-
-    // Then
-    graph should haveConstraints("UNIQUENESS:Person(email)")
-    graph should not(haveConstraints("UNIQUENESS:Person(firstname,lastname)"))
-  }
-
-  test("composite uniqueness constraint should not block adding nodes with different properties") {
-    // When
-    executeWithCostPlannerAndInterpretedRuntimeOnly("CREATE CONSTRAINT ON (n:User) ASSERT (n.firstname,n.lastname) IS UNIQUE")
-
-    // Then
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Soap"), "User")
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Smoke"), "User")
-    createLabeledNode(Map("firstname" -> "Jake", "lastname" -> "Soap"), "User")
-  }
-
-  test("composite uniqueness constraint should block adding nodes with same properties") {
-    // When
-    executeWithCostPlannerAndInterpretedRuntimeOnly("CREATE CONSTRAINT ON (n:User) ASSERT (n.firstname,n.lastname) IS UNIQUE")
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Soap"), "User")
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Smoke"), "User")
-
-    // Then
-    a[ConstraintViolationException] should be thrownBy {
-      createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Soap"), "User")
-    }
-  }
-
-  test("composite uniqueness constraint should not fail when we have nodes with different properties") {
-    // When
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Soap"), "User")
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Smoke"), "User")
-    createLabeledNode(Map("firstname" -> "Jake", "lastname" -> "Soap"), "User")
-
-    // Then
-    executeWithCostPlannerAndInterpretedRuntimeOnly("CREATE CONSTRAINT ON (n:User) ASSERT (n.firstname,n.lastname) IS UNIQUE")
-  }
-
-  test("composite uniqueness constraint should fail when we have nodes with same properties") {
-    // When
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Soap"), "User")
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Smoke"), "User")
-    createLabeledNode(Map("firstname" -> "Joe", "lastname" -> "Soap"), "User")
-
-    // Then
-    a[CypherExecutionException] should be thrownBy {
-      executeWithCostPlannerAndInterpretedRuntimeOnly("CREATE CONSTRAINT ON (n:User) ASSERT (n.firstname,n.lastname) IS UNIQUE")
-    }
-  }
-
   case class haveIndexes(expectedIndexes: String*) extends Matcher[GraphDatabaseQueryService] {
     def apply(graph: GraphDatabaseQueryService): MatchResult = {
       graph.inTx {
@@ -312,20 +252,6 @@ class CompositeIndexAcceptanceTest extends ExecutionEngineFunSuite with NewPlann
           result,
           s"Expected graph to have indexes ${expectedIndexes.mkString(", ")}, but it was ${indexNames.mkString(", ")}",
           s"Expected graph to not have indexes ${expectedIndexes.mkString(", ")}, but it did."
-        )
-      }
-    }
-  }
-
-  case class haveConstraints(expectedConstraints: String*) extends Matcher[GraphDatabaseQueryService] {
-    def apply(graph: GraphDatabaseQueryService): MatchResult = {
-      graph.inTx {
-        val constraintName = graph.schema().getConstraints.asScala.toList.map(i => s"${i.getConstraintType}:${i.getLabel}(${i.getPropertyKeys.asScala.toList.mkString(",")})")
-        val result = expectedConstraints.forall(i => constraintName.contains(i.toString))
-        MatchResult(
-          result,
-          s"Expected graph to have constraints ${expectedConstraints.mkString(", ")}, but it was ${constraintName.mkString(", ")}",
-          s"Expected graph to not have constraints ${expectedConstraints.mkString(", ")}, but it did."
         )
       }
     }
